+++ azext_iot/constants.py
@@ -7,7 +7,7 @@
 
 import os
 
-VERSION = "0.25.0"
+VERSION = "0.26.0"
 EXTENSION_NAME = "azure-iot"
 EXTENSION_ROOT = os.path.dirname(os.path.abspath(__file__))
 EXTENSION_CONFIG_ROOT_KEY = "iotext"
@@ -23,8 +23,8 @@ MESSAGING_HTTP_C2D_SYSTEM_PROPERTIES = [
     "iothub-expiry",
     "iothub-deliverycount",
     "iothub-enqueuedtime",
-    "ContentType",
-    "ContentEncoding",
+    "content-type",
+    "content-encoding",
 ]
 METHOD_INVOKE_MAX_TIMEOUT_SEC = 300
 METHOD_INVOKE_MIN_TIMEOUT_SEC = 10
@@ -47,6 +47,6 @@ IOTHUB_THROTTLE_MAX_TRIES = 3
 IOTHUB_THROTTLE_SLEEP_SEC = 20
 THROTTLE_HTTP_STATUS_CODE = 429
 IOTHUB_RENEW_KEY_BATCH_SIZE = 100
-# (Lib name, minimum version (including), maximum version (excluding))
-EVENT_LIB = ("uamqp", "1.2", "1.3")
+UAMQP_DEP_NAME = "uamqp"
+UAMQP_COMPAT_VERSION = "1.6.6"
 PNP_DTDLV2_COMPONENT_MARKER = "__t"