+++ azext_iot/tests/iothub/core/test_iot_messaging_int.py
@@ -5,13 +5,13 @@
 # --------------------------------------------------------------------------------------------
 
 import os
-from azext_iot.iothub.common import NON_DECODABLE_PAYLOAD
-from azext_iot.tests.conftest import get_context_path
 import pytest
 import json
+from time import time, sleep
 
-from time import time
 from uuid import uuid4
+from azext_iot.iothub.common import NON_DECODABLE_PAYLOAD
+from azext_iot.tests.conftest import get_context_path
 from azext_iot.tests.helpers import CERT_ENDING, KEY_ENDING
 from azext_iot.tests.iothub import IoTLiveScenarioTest, PREFIX_DEVICE
 from azext_iot.common.utility import (
@@ -101,8 +101,8 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
         assert result["data"] == test_body
 
         system_props = result["properties"]["system"]
-        assert system_props["ContentEncoding"] == test_ce
-        assert system_props["ContentType"] == test_ct
+        assert system_props["content-encoding"] == test_ce
+        assert system_props["content-type"] == test_ct
         assert system_props["iothub-correlationid"] == test_cid
         assert system_props["iothub-messageid"] == test_mid
         assert system_props["iothub-expiry"]
@@ -157,8 +157,8 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
             self._remove_newlines_spaces(payload=self.kwargs["messaging_data"])
 
         system_props = result["properties"]["system"]
-        assert system_props["ContentEncoding"] == test_ce
-        assert system_props["ContentType"] == 'application/json'
+        assert system_props["content-encoding"] == test_ce
+        assert system_props["content-type"] == 'application/json'
         assert system_props["iothub-correlationid"] == test_cid
         assert system_props["iothub-messageid"] == test_mid
         assert system_props["iothub-expiry"]
@@ -209,8 +209,8 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
         assert result["data"] == self.kwargs["messaging_unicodable_data"]
 
         system_props = result["properties"]["system"]
-        assert system_props["ContentEncoding"] == test_ce
-        assert system_props["ContentType"] == 'application/octet-stream'
+        assert system_props["content-encoding"] == test_ce
+        assert system_props["content-type"] == 'application/octet-stream'
         assert system_props["iothub-correlationid"] == test_cid
         assert system_props["iothub-messageid"] == test_mid
         assert system_props["iothub-expiry"]
@@ -261,8 +261,8 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
         assert result["data"] == self.kwargs["messaging_non_unicodable_data"]
 
         system_props = result["properties"]["system"]
-        assert system_props["ContentEncoding"] == test_ce
-        assert system_props["ContentType"] == 'application/octet-stream'
+        assert system_props["content-encoding"] == test_ce
+        assert system_props["content-type"] == 'application/octet-stream'
         assert system_props["iothub-correlationid"] == test_cid
         assert system_props["iothub-messageid"] == test_mid
         assert system_props["iothub-expiry"]
@@ -310,11 +310,12 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
             )
         ).get_output_in_json()
 
-        assert result["data"] == self.kwargs["messaging_non_unicodable_data"]
+        # no data in this result
+        # assert result["data"] == self.kwargs["messaging_non_unicodable_data"]
 
         system_props = result["properties"]["system"]
-        assert system_props["ContentEncoding"] == 'gzip'
-        assert system_props["ContentType"] == 'application/octet-stream'
+        assert system_props["content-encoding"] == 'gzip'
+        assert system_props["content-type"] == 'application/octet-stream'
         assert system_props["iothub-correlationid"] == test_cid
         assert system_props["iothub-messageid"] == test_mid
         assert system_props["iothub-expiry"]
@@ -386,8 +387,8 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
         assert result["data"] == self.kwargs["c2d_json_send_data"]
 
         system_props = result["properties"]["system"]
-        assert system_props["ContentEncoding"] == test_ce
-        assert system_props["ContentType"] == test_ct
+        assert system_props["content-encoding"] == test_ce
+        assert system_props["content-type"] == test_ct
         assert system_props["iothub-correlationid"] == test_cid
         assert system_props["iothub-messageid"] == test_mid
         assert system_props["iothub-expiry"]
@@ -771,8 +772,11 @@ class TestIoTHubMessaging(IoTLiveScenarioTest):
 
         # x509 CA device simulation and include model Id upon connection
         model_id_simulate_x509ca = "dtmi:com:example:simulatex509ca;1"
+
+        # not sure why this needs a timer but it seems to help avoid unauthorized errors
+        sleep(60)
         self.cmd(
-            "iot device simulate -d {} -n {} -g {} --da '{}' --mc 1 --mi 1 --cp {} --kp {} --pass {} --model-id {}".format(
+            "iot device simulate -d {} -n {} -g {} --da '{}' --mc 1 --mi 1 --cp {} --kp {} --pass {} --model-id '{}'".format(
                 device_ids[1], self.entity_name, self.entity_rg, simulate_msg,
                 f"{device_ids[1]}-cert.pem", f"{device_ids[1]}-key.pem", fake_pass, model_id_simulate_x509ca
             )